pyproject.toml
[project]
name = "external-batches-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "resim-sdk>=1.1.0",
]
